pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "openai>=1.76.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
